fcoemon: netlink buffer resize fix
The recv buffer resizing for netlink messages looks to have always been
busted, it discards some of the data in the process of increasing the
buffer size.

I've recently seen issues where the size of the netlink attributes in a
GETLINK request hit the fcm_link_buf_size window and some of the network
interfaces were ignored by fcoemon (happened mostly with recent kernels,
2 dual-port bnx2 NICs, bonding for LAN traffic on 2 ports and FCoE on
the other 2).  When the ignored interface happens to be the one you want
to use for FCoE, it never gets fixed.

This fixes the buffer resize code to use MSG_PEEK|MSG_TRUNC to check for
the needed buffer size first, and then resize with realloc when needed.
This might not actually be needed, with an 8k buffer I don't think the
kernel side will send more at once until after it sees the application
post a larger buffer, but I did force a resize and see that it worked.

Signed-off-by: Chris Leech <cleech@redhat.com>