fix regressions caused by safe_makepath change in libopenfcoe.c
This needs to use strlen(dname) not sizeof(dname) or the directory path
is truncated.  On a 64-bit arch an 8 byte truncation kind of looks like
a valid path '/sys/bus' which is sort of funny unless you're debugging
it.

This caused fcoeadm --fcf to fail to read in FCF info, and fcoeadm
--lesb to fail to find the link error status block counters.

Signed-off-by: Chris Leech <cleech@redhat.com>