fcoemon: add snprintf string precision modifiers in fcm_netif_advance
GCC 12 is warning of potential snprintf truncations

fcm_netif.ifname is an IFNAMSIZ array, but formating with %s doesn't
understand that, so add a precision modifier every time we print it to
limit the output.  This allows the compiler to verify that the output
buffer is of sufficient length to never truncate.

Signed-off-by: Chris Leech <cleech@redhat.com>